backend/requirements.txt
langchain>=0.1.0
langgraph>=0.0.20
@arkiv-network/sdk==1.0.0a5
replicate>=0.15.0
openai>=1.10.0
web3>=6.15.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pytest>=7.4.0
pyyaml>=6.0.1
aiofiles>=23.2.1

